Improve logging in 'Deploy to ECR' stage of Jenkinsfile
The `BRANCH_NAME` variable is now printed immediately after it is defined, aiding in debugging by showing the branch name in the console output. The `commitHash` variable assignment has been moved to a new line after the `CLEAN_NAME` variable is printed, ensuring that the branch name and its cleaned version are logged before the commit hash is retrieved. These changes enhance the logging and debugging capabilities of the deployment script.